==> computational_notebooks/02-metadata_additional_tags.ipynb <==
# Data Analytics: Exploring Specific Tags

Cara Marta Messina <br/>
Northeastern University<br/>
messina [dot] c [at] husky [dot] neu [dot] edu

This notebook takes data collected from <em>Archive of Our Own</em>, a popular fanfiction repository, and sets it up to be analyzed. The data was collected using [this AO3 python scraper](https://github.com/radiolarian/AO3Scraper). The corpus consists of <em>The Legend of Korra</em> and <em>Game of Thrones</em> fanfics, from the first one published on AO3 to 2019.

<em>This notebook is part of the Critical Fan Toolkit, Cara Marta Messina's public + digital dissertation</em>

#pandas for working with dataframes
import pandas as pd

#regular expression library
import re

#numpy specifically works with numbers
import numpy as np

from nltk import word_tokenize

import string
punctuations = list(string.punctuation)

#has the nice counter feature for counting tags
import collections
from collections import Counter 

#for making a string of elements separated by commas into a list
from nltk.tokenize.punkt import PunktSentenceTokenizer, PunktLanguageVars 

#visualizations
#import plotly as py
#import plotly.express as px
#import plotly.graph_objs as go
#from plotly.subplots import make_subplots

#calling my plotly thing
#import chart_studio
#chart_studio.tools.set_credentials_file(username='caramessina', api_key='IdA4LjtaqYKmFJnfS8Uv')

## Reading and Prepping the Data

korra_all = pd.read_csv('./data/allkorra.csv')
korra_all.head(3)

#reading in multiple csv files, since one large one breaks my kernels 

got0 = pd.read_csv(r'./data/got_data_clean/got0.csv')
got1 = pd.read_csv(r'./data/got_data_clean/got1.csv')
got2 = pd.read_csv(r'./data/got_data_clean/got2.csv')
got3 = pd.read_csv(r'./data/got_data_clean/got3.csv')

merged_got = pd.concat([got0, got1, got2, got3])
merged_got.head(3)

merged_got.count()

## Creating the Functions: Extracting Metadata

created a function that will take the different tags (which are phrased as characterA/characterB, characterA/characterB, etc in the data) and count the most common relationships to then output it as the most common relationship tags used. 

def column_to_list(df,columnName):
    '''
    this function takes all the information from a specific column, joins it to a string, and then tokenizes & cleans that string.
    input: the name of the dataframe and the column name
    output: the tokenized list of the text with all lower case, punctuation removed, and no stop words
    '''
    df[columnName] = df[columnName].replace(np.nan,'',regex=True) 
    string = ' '.join(df[columnName].tolist())
    return string

def clean_tokens(string):    
    stopwords = ['i', 'me', 'my', 'myself', "“", "”", 'we', 'our', '’', 'ours', 'ourselves', 'you', "you're", "you've", "you'll", "you'd", 'your', 'yours', 'yourself', 'yourselves', 'he', 'him', 'his', 'himself', 'she', "she's", 'her', 'hers', 'herself', 'it', "it's", 'its', 'itself', 'they', 'them', 'their', 'theirs', 'themselves', 'what', 'which', 'who', 'whom', 'this', 'that', "that'll", 'these', 'those', 'am', 'is', 'are', 'was', 'were', 'be', 'been', 'being', 'have', 'has', 'had', 'having', 'do', 'does', 'did', 'doing', 'a', 'an', 'the', 'and', 'but', 'if', 'or', 'because', 'as', 'until', 'while', 'of', 'at', 'by', 'for', 'with', 'about', 'against', 'between', 'into', 'through', 'during', 'before', 'after', 'above', 'below', 'to', 'from', 'up', 'down', 'in', 'out', 'on', 'off', 'over', 'under', 'again', 'further', 'then', 'once', 'here', 'there', 'when', 'where', 'why', 'how', 'all', 'any', 'both', 'each', 'few', 'more', 'most', 'other', 'some', 'such', 'no', 'nor', 'not', 'only', 'own', 'same', 'so', 'than', 'too', 'very', 's', 't', 'can', 'will', 'just', 'don', "don't", 'should', "should've", 'now', 'd', 'll', 'm', 'o', 're', 've', 'y', 'ain', 'aren', "aren't", 'couldn', "couldn't", 'didn', "didn't", 'doesn', "doesn't", 'hadn', "hadn't", 'hasn', "hasn't", 'haven', "haven't", 'isn', "isn't", 'ma', 'mightn', "mightn't", 'mustn', "mustn't", 'needn', "needn't", 'shan', "shan't", 'shouldn', "shouldn't", 'wasn', "wasn't", 'weren', "weren't", "would", "could", 'won', "won't", 'wouldn', "wouldn't"]
    text_lc = [word.lower() for word in string]
    text_tokens_clean = [word for word in text_lc if word not in stopwords]
    text_tokens_clean = [word for word in text_tokens_clean if word not in punctuations]
    return text_tokens_clean
    print(text_tokens_clean[:20])

def TagsAnalyzer(df, columnName):
    '''
    input: the index month+year, such as '2012-04', and the specific metadata, such as 'additional tags'
    output: a tupple of the count of tags in a specific month/year
    
    load in the proper data into a string'''
    
    #replace empty values & make a list of all the words
    string = column_to_list(df, columnName)
    
    #the function to tokenize, or put each value as an element in a list
    class CommaPoint(PunktLanguageVars):
        sent_end_chars = (',') 
    tokenizer = PunktSentenceTokenizer(lang_vars = CommaPoint())
    
    #tokenizing the list of strings based on the COMMA, not the white space (as seen in the CommaPoint above)
    ListOfTags = tokenizer.tokenize(string)
    
    #the "Counter" function is from the collections library
    allCounter=collections.Counter(ListOfTags)
    
    li = []
    #returning a dictionary in which the keys are all the tags, and the items are the counts
    for word0,word1 in allCounter.most_common(300):
        li.append((word0, (word1/length)*100))

#     #return
    return li

def extract_tags(dictionary,search_term,count):
    '''
    input: a dictionary to be searched, the term you want to search, and the total number of tags or whatever you count
    output: all the items in which your specific term appears and the ratio of times each one appears
    '''
    for tag, number in dictionary.items():    # for name, age in dictionary.iteritems():  (for Python 2.x)
        if search_term in tag:
            print(tag, number, (number/count)*100)

## The Legend of Korra: Exploring Additional Tags

I've already done some of this work, but I want to add more potential writers to the list. What is really interesting about these tags, though, is there are many less of them. I wonder, because Korra/Asami is the "norm", if there needs to be a labeling of difference. For example, Korrasami/KorraxKuvira is much more used than other relationships in this fandom, so being explicitly open about its queerness isn't important. Tags don't need to be used to define race/queerness because these identities are already so embedded in both the show and the fanfics.

For example, feminism and pointing to feminism does <em>not</em> really happen.

### Race
- canon character of color
- male character of color
- character of color
- lgbtq female characters of color
- 'of color' is a tag to look at

### Queer/Lesbian/Gay/Gender
- socially awkward queer girls
- genderqueer korra
- okay it would be smut but i started writing a lesbian adventure romcom
- kya being a horribly flirtatious lesbian
- nonbinary lesbians
- gaycation all i ever wanted
- everyone's gay/everyone is gay
- idk how else to describe this but gay gay gay
- older gays giving advice to younger gays
- drag queen/drag queens
- trans
- trans female character
- trans characters
- trans!korraxasami
- pre-transition
- for trans!korra to help dysphoria the other way around
- genderqueer 
- gender dysphoria/euphoria
- gender queer characters

tlok_AT_count = korra_all['additional tags'].str.len().sum()

#All of the Additional Tags used in the Game of Thrones fandom
tlok_AT = TagsAnalyzer(korra_all,'additional tags')

extract_tags(tlok_AT,'color',tlok_AT_count)
extract_tags(tlok_AT,'race',tlok_AT_count) #...master race?
extract_tags(tlok_AT,'queer',tlok_AT_count)
extract_tags(tlok_AT,'lesbian',tlok_AT_count)
extract_tags(tlok_AT,'gay',tlok_AT_count)
extract_tags(tlok_AT,'drag',tlok_AT_count)
extract_tags(tlok_AT,'bisex',tlok_AT_count)
extract_tags(tlok_AT,'trans',tlok_AT_count)
extract_tags(tlok_AT,'feminis',tlok_AT_count) #less feminist tags coming up–FASCINATING
extract_tags(tlok_AT,'gender',tlok_AT_count)

## Game of Thrones: Exploring Additional Tags

In order to better discover <em>Game of Thrones</em> fanfics that are thinking about race, sexuality, and gender in nuanced ways, I will first search for particular tags that are used and create a list of tags here that I will be using to then chose the particular fics I want to read.

### Race
- female character of color
- character of color
- <strong>pov character of color</strong>
- <strong>bisexual female character of color</strong>
- <strong>canon character of color</strong>
- <strong>canon queer character of color</strong>
- lesbian character of color
- lgbtq character(s) of color
- <strong>male character of color</strong>
- characters of color
- lgbtq character of color
- <strong>race relations</strong>
- <strong>racebending</strong>
- race
- alternate universe - race changes
- <strong>reference to race relations</strong>
- different races (not interesting)

### Racism
- fantastic racism
- racism
- period-typical racism
- magical racism (not interesting)
- racism allegory
- just a little bit of valyrian racism

### Gender
- gender issues
- genderfuck
- <strong>genderfluid character</strong>
- gender roles
- gender equality
- gender confusion
- transgender
- gender identity
- arya is genderqueer
- gender neutrality
- gender dysphoria
- <strong>oberyn gives zero fucks about your gendered shit</strong>
- gender-neutral pronouns
- genderqueer jon
- <strong>masculine presenting gender neutral jon</strong>
- transgender teens
- <strong>fuck gender roles</strong>

### Queer
- queer character
- <strong>original character is an autistic queer woman of colour</strong>
- cersei and jaime are genderqueer
- queer
- queer themes
- <strong>canon queer character of color</strong>
- canon queer character
- pov queer character
- everyone is queer
- arya is queer
- <strong>almost everyone is queer</strong>
- queer characters
- queer youth
- queer identity
- queer friendships
- canon queer relationship
- queer queens
- bisexual jaime is here and queer
- queer daenerys
- queer culture
- <strong>trans character</strong>
- transphobic language
- canon lgbtq female character
- <strong>jon is a trans man please stop telling me to tag this 'female jon snow'</strong>
- jon snow is trans
- <strong>alternate universe - trans</strong>
- trans! margaery
- trans!jon snow
- <strong>trans!tywin lannister</strong>
- trans daenerys
- i edited it to make it trans inclusive
- trans female character
- <strong>trans!arya stark</strong>
- trans loras

### Lesbian
There are a lot of results here. I think just keeping "lesbian" and "queer" as basic search terms already pulls up what I need it to.

### Fem/Masc (and feminism)
- <strong>feminist themes</strong>
- female sexuality
- female-centric
- strong female characters
- bisexual female character
- trans female character
- lack of female rights
- <strong>calling the show out on its overuse of female nudity</strong>
- <strong>there's not enough femslash here</strong>
- female jon snow is also autistic
- feminist wives everywhere
- <strong>the idealization of violence makes for shitty feminism</strong>
- feminism
- <strong>my response to 'female characters can only have growth through physical pain</strong>
- feminism. idk.
- <strong>because fuck d&d and their offending representation of female relationships</strong>
- <strong>au in which female characters have agency</strong>
- <strong>tyrion lannister is a feminist icon</strong>
- <strong>robb stark doesn't have time for your hypermasculine bullshit</strong>
- toxic masculinity
- <strong>tormund is my feminist</strong>

### Consent
- <strong>explicit description of consent because that's important y'all</strong>
- <strong>explicit consent</strong>
- consensual non-consent
- enthusiastic consent
- so much consent
- dubious consent due to identity issues
- consentual sex
- consenting adults
- consent is v important folks
- consent is sexy
- <strong>look d&d you can't consent when you're that drunk</strong>
- <strong>jaime has a consent kink</strong>
- consent is sexy kids
- but everyone consents
- comment on consent (hmm what is this one)

### Loras
- loras' pov
- fem!loras
- <strong>loras is king of coming out parties</strong>
- pov loras tyrell
- <strong>loras doesn't get the recognition he deserves</strong>

### Missandei/Grey Worm
- <strong>missandei deserved better</strong>
- <strong>bisexual missandei</strong>
- daenerys targaryen/missandei
- <strong>inspired by grey worm and missandei goodbye scene</strong>
- <strong>dany's army now waiting for missandei's word</strong>
- <strong>i haven't coped with missandei's death</strong>
- <strong>a fix it fic because its what missandei deserves goddamnit</strong>
- missandei is an angel
- missandei is a precious angel who must be protected
- daenerys and missandei have to say goodbye
- brotp arya & missandei (instead of OTP, it's a friendship pairing)
- <strong>grey worm deserved better</strong>
- <strong>y'all need to write stories about the trauma grey worm is going through</strong>

### Misc
- <strong>slavery abolition</strong>
- non-romanticized slavery – not GOT
- there is no canon to diverge from anymore-freedom is now (no idea what this means, but interested)

merged_got.count()

got_AT_count = merged_got['additional tags'].str.len().sum()

#All of the Additional Tags used in the Game of Thrones fandom
got_AT = TagsAnalyzer(merged_got,'additional tags')

extract_tags(got_AT,'color',got_AT_count)
extract_tags(got_AT,'race',got_AT_count)
extract_tags(got_AT,'racism',got_AT_count)
extract_tags(got_AT,'black people',got_AT_count)
extract_tags(got_AT,'gender',got_AT_count)
extract_tags(got_AT,'queer',got_AT_count)
extract_tags(got_AT,'lesbian',got_AT_count)
extract_tags(got_AT,'gay',got_AT_count)
extract_tags(got_AT,'nonbinary',got_AT_count)
extract_tags(got_AT,'genderqueer',got_AT_count)
extract_tags(got_AT,'missandei',got_AT_count)
extract_tags(got_AT,'grey worm',got_AT_count)
extract_tags(got_AT,'dothraki',got_AT_count)
extract_tags(got_AT,'khal',got_AT_count)
extract_tags(got_AT,'dysphoria',got_AT_count)
extract_tags(got_AT,'slave',got_AT_count)
extract_tags(got_AT,'drag ',got_AT_count)
extract_tags(got_AT,'abolition',got_AT_count)
extract_tags(got_AT,'freedom',got_AT_count)
extract_tags(got_AT,'revolution',got_AT_count)
extract_tags(got_AT,'rebellion',got_AT_count)
extract_tags(got_AT,"y'all",got_AT_count)
extract_tags(got_AT,"consent",got_AT_count)
extract_tags(got_AT,"trans",got_AT_count)
extract_tags(got_AT,"loras",got_AT_count)
extract_tags(got_AT,"fem",got_AT_count)
extract_tags(got_AT,"masc",got_AT_count)
extract_tags(got_AT,"feminis",got_AT_count)
extract_tags(got_AT,"heterono",got_AT_count)
extract_tags(got_AT,"autistic",got_AT_count)

## Extracting Fics that Use These Tags

Using the tag list above, I will extract fics that use these tags to determine if I want to read the fics and reach out to the authors!

Instead of putting all this information in the Jupyer Notebook, though, I will be putting it all in a Google Sheet where I will be keeping track of all this information

These are the tags I am choosing to extract:
- Missandei
- Grey Worm
- feminis (femism, femist)
    - the idealization of violence makes for shitty feminism
    - feminist themes
    - feminism
- trans!
- nonbinary
    - nonbinary character
    - nonbinary arya
- transge
- femal
    - my response to 'female characters can only have growth through physical pain
    - female character in command
    - female friendships yo
- gay
- bisex
- LGBTQ
- dothraki
- heteronormativity
- lesbian
- dysphoria
    - body dysphoria
- gender
    - genderqueer
    - genderqueer jon
    - genderqueer character
    - genderfluid arya
    - gender role expectations
    - oberyn gives zero fucks about your gendered shit
    - gender confusion
    - genderbending
    - misgenderin
- abolition
    - slavery abolition
- of color
    - canon character of color
    - character(s) of color
    - lgbtq character(s) of color
- of colour
- masc
- queer
- ' race' (put the space before race so it's not a word like 'trace')
    - racebend 
    - 
- drag (specifically "drag " with a space at the end to eliminate dragons)
    - drag queen

Decided not to include:
- racism, as this is usually a justification to include racism, claiming it is "canonical" or part of the world
- consent, because most things that mention consent use the tag 'dubious consent.' There are some tags that emphasize the need for consent, but these are much more rare.
- slash, as often slash is centered around two male characters, both erasing women and fetishizing gay men


def find_text(search_term):
    '''
    Simple function so I dont need to copy-and-paste over and over
    '''
    show = got_all[got_all['additional tags'].str.contains(search_term)]
    print(show)

Got_LGBTQ = merged_got[merged_got['additional tags'].str.contains("slash|lgbtq|feminis|drag |nonbinary|bisex|heteronormativity|trans!|transg|dysphoria|gay| bi |lesbian|genderqu|gender iden|gender neu|masc|queer", na=False)]
Got_LGBTQ.count()

GoT_race_AT = merged_got[merged_got['additional tags'].str.contains("missandei|grey worm|dothraki|abolition|of color|of colour| race|naath|meereen|oberyn|ellaria|dorne|coloni", na=False)]
GoT_race_AT.count()

def character_count(charactername,df):
    df_new = df[df['character'].str.contains(charactername, na=False)]
    print(charactername, df_new['work_id'].count())

character_count('jon snow', merged_got)
character_count('tyrion', merged_got)
character_count('arya', merged_got)
character_count('sansa stark', merged_got)
character_count('sandor', merged_got)
character_count('baelish', merged_got)
character_count('daenerys targaryen', merged_got)
character_count('brienne', merged_got)
character_count('margaery', merged_got)
character_count('jaime', merged_got)
character_count('loras', merged_got)
character_count('oberyn', merged_got)
character_count('missandei', merged_got)
character_count('grey worm', merged_got)
character_count('ellaria', merged_got)
character_count('bran', merged_got)
character_count('drogo', merged_got)

GoT_race_REL = merged_got[merged_got['relationship'].str.contains("missandei|grey worm|khal|jhiqui|irri|qotho|rakharo|sallador saan|xaro|oberyn|ellaria", na=False)]
GoT_race_REL.count()

GoT_race_REL = merged_got[~merged_got['relationship'].str.contains("missandei|grey worm|khal|jhiqui|irri|qotho|rakharo|sallador saan|xaro|oberyn|ellaria", na=False)]
GoT_race_REL.count()

GoT_chr_REL = merged_got[merged_got['character'].str.contains("missandei|grey worm|khal|jhiqui|irri|qotho|rakharo|sallador saan|xaro|oberyn|ellaria", na=False)]
GoT_chr_REL.count()

#did not include slavery because sometimes slavery is used in a sexual connotation, not in the historic and politic and violent way

GOT_woke = merged_got[merged_got['additional tags'].str.contains("lgbtq|missandei|grey worm|femal|feminis|drag |bisex|heteronormativity|trans!|transg|dysphoria|consent|gay|nonbinary|dothraki|lesbian|gender|abolition|of color|of colour|naath|meereen|oberyn|ellaria|dorne|coloni|masc|queer| race", na=False)]
GOT_woke

GOT_maybeWokeIDK = merged_got[merged_got['additional tags'].str.contains("lgbtq|missandei|grey worm|femal|feminis|drag |bisex|heteronormativity|trans!|transg|dysphoria|consent|gay|dothraki|lesbian|gender|abolition|of color|of colour|naath|meereen|oberyn|ellaria|dorne|coloni|masc|queer| race")==False]
GOT_maybeWokeIDK

korra_all[korra_all['additional tags'].str.contains("rans!",  na=False)]
